@@ -8,18 +8,18 @@ import 'package:isolate_bloc/src/common/isolate/service_events.dart';
 import '../bloc/isolate_bloc.dart';
 
 /// Signature for function which creates [IsolateBloc].
-typedef IsolateBlocCreator<Event, State> = IsolateBloc<Event, State> Function();
+typedef IsolateBlocCreator<Event extends Object, State extends Object>
+    = IsolateBloc<Event, State> Function();
 
 /// Maintain [IsolateBloc]s in isolate
 class IsolatedBlocManager {
   IsolatedBlocManager._(this._isolatedConnector);
 
-  static IsolatedBlocManager instance;
+  static IsolatedBlocManager? instance;
   final IsolatedConnector _isolatedConnector;
   final _initialStates = <Type, Object>{};
   final _freeBlocs = <Type, IsolateBloc>{};
   final _blocCreators = <Type, IsolateBlocCreator>{};
-  final _registeredWrappers = <Type, Object>{};
   final _createdBlocs = <String, IsolateBloc>{};
   final _initializeCompleter = Completer();
 
@@ -35,27 +35,42 @@ class IsolatedBlocManager {
   }
 
   /// Returns new bloc from cached in [_freeBlocs] or create new one.
-  IsolateBloc _getFreeBlocByType(Type type) {
-    IsolateBloc bloc;
+  IsolateBloc? _getFreeBlocByType(Type type) {
+    IsolateBloc? bloc;
     if (_freeBlocs.containsKey(type)) {
-      bloc = _freeBlocs.remove(type);
+      bloc = _freeBlocs.remove(type)!;
     } else {
-      bloc = _blocCreators[type]();
+      final blocCreator = _blocCreators[type];
+      if (blocCreator == null) {
+        print("Can't create IsolateBloc with type $type.\n"
+            "Maybe you forgot to register it?");
+      } else {
+        bloc = blocCreator.call();
+      }
     }
-    _createdBlocs[bloc.id] = bloc;
-    return bloc;
+    if (bloc != null) {
+      _createdBlocs[bloc.id] = bloc;
+      return bloc;
+    } else {
+      return null;
+    }
   }
 
-  Future<T> _getBloc<T extends IsolateBloc>() async {
+  Future<T?> _getBloc<T extends IsolateBloc>() async {
     await _initializeCompleter.future;
-    // find every blocs with Type T
     final blocsT = _createdBlocs.values.whereType<T>();
     if (blocsT.isNotEmpty) {
       return blocsT.first;
     } else if (_freeBlocs.containsKey(T)) {
       return _freeBlocs[T] as T;
     } else {
-      return (_freeBlocs[T] = _blocCreators[T]()) as T;
+      final blocCreator = _blocCreators[T];
+      if (blocCreator == null) {
+        print("Failed to find BlocCreator for $T");
+        return null;
+      } else {
+        return _freeBlocs[T] = blocCreator() as T;
+      }
     }
   }
 
@@ -66,14 +81,18 @@ class IsolatedBlocManager {
   /// returns this bloc's [IsolateBlocWrapper]. Else it is creates a new bloc and
   /// add to the pull of free blocs. So when UI will call `create()`, it will not create a new bloc but
   /// return free bloc from pull.
-  IsolateBlocWrapper<State>
-      getBlocWrapper<Bloc extends IsolateBloc<Object, State>, State>() {
-    IsolateBlocWrapper<State> wrapper;
-    Bloc isolateBloc;
+  IsolateBlocWrapper<State> getBlocWrapper<
+      Bloc extends IsolateBloc<Object, State>, State extends Object>() {
+    late IsolateBlocWrapper<State> wrapper;
+    Bloc? isolateBloc;
     _getBloc<Bloc>().then((bloc) {
-      isolateBloc = bloc;
-      bloc.listen(wrapper.stateReceiver);
-      wrapper.connectToBloc(bloc.id);
+      if (bloc == null) {
+        print("Failed to get bloc with type $Bloc");
+      } else {
+        isolateBloc = bloc;
+        bloc.listen(wrapper.stateReceiver);
+        wrapper.connectToBloc(bloc.id);
+      }
     });
     final onBLocClose = (_) => isolateBloc?.close();
     final eventReceiver = (IsolateBlocTransitionEvent<Object> event) =>
@@ -85,19 +104,18 @@ class IsolatedBlocManager {
   /// Receive bloc's [uuid] and [event].
   /// Find [IsolateBloc] by id and add [event] to it.
   void blocEventReceiver(String uuid, Object event) {
-    _createdBlocs[uuid].add(event);
+    final bloc = _createdBlocs[uuid];
+    if (bloc == null) {
+      print("Failed to receive event. Bloc doesn't exist");
+    } else {
+      bloc.add(event);
+    }
   }
 
   /// Create [IsolateBloc] and connect it to the [IsolateBlocWrapper].
   void createBloc(Type blocType) {
-    assert(
-      _blocCreators.containsKey(blocType) ||
-          _registeredWrappers.containsKey(blocType),
-      'You must register bloc or bloc wrapper to create it.',
-    );
-    if (_blocCreators.containsKey(blocType)) {
-      // ignore: close_sinks
-      final bloc = _getFreeBlocByType(blocType);
+    final bloc = _getFreeBlocByType(blocType);
+    if (bloc != null) {
       bloc.listen(
         (state) => _isolatedConnector.sendEvent(
           IsolateBlocTransitionEvent(bloc.id, state),
@@ -105,26 +123,23 @@ class IsolatedBlocManager {
       );
       _isolatedConnector
           .sendEvent(IsolateBlocCreatedEvent(bloc.runtimeType, bloc.id));
-    } else {
-      BlocManager.instance.createBloc();
     }
   }
 
   /// Get bloc by [uuid] and close it.
   void closeBloc(String uuid) {
-    assert(
-      _createdBlocs.containsKey(uuid),
-      'You are trying to remove bloc with with a nonexistent uuid($uuid).\n'
-      'This can happen if you call close() twice.',
-    );
-    _createdBlocs.remove(uuid).close();
+    final bloc = _createdBlocs.remove(uuid);
+    if (bloc == null) {
+      print("Failed to close bloc because it wasn't created yet.");
+    } else {
+      bloc.close();
+    }
   }
 
   /// Register [IsolateBloc].
   /// You can create [IsolateBloc] and get [IsolateBlocWrapper] from
   /// [BlocManager].createBloc only if you register this [IsolateBloc].
-  void register<Event, State>(IsolateBlocCreator<Event, State> creator) {
-    // ignore: close_sinks
+  void register<Event, State>(IsolateBlocCreator creator) {
     final bloc = creator();
     _initialStates[bloc.runtimeType] = bloc.state;
     _freeBlocs[bloc.runtimeType] = bloc;
