@@ -1,10 +1,8 @@
-import 'package:flutter/foundation.dart';
-
 /// Library info for [PlatformChannelSetup].
 class Library {
   const Library({
-    @required this.name,
-    @required this.methodChannels,
+    required this.name,
+    required this.methodChannels,
   });
 
   final String name;
