@@ -14,7 +14,9 @@ mixin IsolateBlocListenerSingleChildWidget on SingleChildWidget {}
 /// with the `state` and is responsible for executing in response to
 /// `state` changes.
 typedef IsolateBlocWidgetListener<S> = void Function(
-    BuildContext context, S state);
+  BuildContext context,
+  S state,
+);
 
 /// Signature for the `listenWhen` function which takes the previous `state`
 /// and the current `state` and is responsible for returning a [bool] which
@@ -79,29 +81,23 @@ typedef IsolateBlocListenerCondition<S> = bool Function(S previous, S current);
 /// )
 /// ```
 /// {@endtemplate}
-class IsolateBlocListener<C extends IsolateBloc<Object, S>, S>
+class IsolateBlocListener<C extends IsolateBloc<Object, S>, S extends Object>
     extends IsolateBlocListenerBase<C, S>
     with IsolateBlocListenerSingleChildWidget {
   /// {@macro bloc_listener}
   const IsolateBlocListener({
-    Key key,
-    @required IsolateBlocWidgetListener<S> listener,
-    IsolateBlocWrapper<S> isolateBloc,
-    IsolateBlocListenerCondition<S> listenWhen,
-    this.child,
-  })  : assert(listener != null),
-        super(
+    Key? key,
+    required IsolateBlocWidgetListener<S> listener,
+    IsolateBlocWrapper<S>? isolateBloc,
+    IsolateBlocListenerCondition<S>? listenWhen,
+    Widget? child,
+  }) : super(
           key: key,
           child: child,
           listener: listener,
           isolateBloc: isolateBloc,
           listenWhen: listenWhen,
         );
-
-  /// The widget which will be rendered as a descendant of the [IsolateBlocListener].
-  @override
-  // ignore: overridden_fields
-  final Widget child;
 }
 
 /// {@template bloc_listener_base}
@@ -111,11 +107,11 @@ class IsolateBlocListener<C extends IsolateBloc<Object, S>, S>
 /// The type of the state and what happens with each state change
 /// is defined by sub-classes.
 /// {@endtemplate}
-abstract class IsolateBlocListenerBase<C extends IsolateBloc<Object, S>, S>
-    extends SingleChildStatefulWidget {
+abstract class IsolateBlocListenerBase<C extends IsolateBloc<Object, S>,
+    S extends Object> extends SingleChildStatefulWidget {
   /// {@macro bloc_listener_base}
   const IsolateBlocListenerBase({
-    Key key,
+    Key? key,
     this.listener,
     this.isolateBloc,
     this.child,
@@ -124,30 +120,30 @@ abstract class IsolateBlocListenerBase<C extends IsolateBloc<Object, S>, S>
 
   /// The widget which will be rendered as a descendant of the
   /// [IsolateBlocListenerBase].
-  final Widget child;
+  final Widget? child;
 
   /// The [isolateBloc] whose `state` will be listened to.
   /// Whenever the [isolateBloc]'s `state` changes, [listener] will be invoked.
-  final IsolateBlocWrapper<S> isolateBloc;
+  final IsolateBlocWrapper<S>? isolateBloc;
 
   /// The [IsolateBlocWidgetListener] which will be called on every `state` change.
   /// This [listener] should be used for any code which needs to execute
   /// in response to a `state` change.
-  final IsolateBlocWidgetListener<S> listener;
+  final IsolateBlocWidgetListener<S>? listener;
 
   /// {@macro bloc_listener_listen_when}
-  final IsolateBlocListenerCondition<S> listenWhen;
+  final IsolateBlocListenerCondition<S>? listenWhen;
 
   @override
   SingleChildState<IsolateBlocListenerBase<C, S>> createState() =>
       _BlocListenerBaseState<C, S>();
 }
 
-class _BlocListenerBaseState<C extends IsolateBloc<Object, S>, S>
+class _BlocListenerBaseState<C extends IsolateBloc<Object, S>, S extends Object>
     extends SingleChildState<IsolateBlocListenerBase<C, S>> {
-  StreamSubscription<S> _subscription;
-  S _previousState;
-  IsolateBlocWrapper<S> _isolateBlocWrapper;
+  StreamSubscription<S>? _subscription;
+  S? _previousState;
+  IsolateBlocWrapper<S>? _isolateBlocWrapper;
 
   @override
   void initState() {
@@ -175,7 +171,9 @@ class _BlocListenerBaseState<C extends IsolateBloc<Object, S>, S>
   }
 
   @override
-  Widget buildWithChild(BuildContext context, Widget child) => child;
+  Widget buildWithChild(BuildContext context, Widget? child) {
+    return child ?? const SizedBox.shrink();
+  }
 
   @override
   void dispose() {
@@ -185,9 +183,9 @@ class _BlocListenerBaseState<C extends IsolateBloc<Object, S>, S>
 
   void _subscribe() {
     if (_isolateBlocWrapper != null) {
-      _subscription = _isolateBlocWrapper.listen((state) {
-        if (widget.listenWhen?.call(_previousState, state) ?? true) {
-          widget.listener(context, state);
+      _subscription = _isolateBlocWrapper?.listen((state) {
+        if (widget.listenWhen?.call(_previousState!, state) ?? true) {
+          widget.listener?.call(context, state);
         }
         _previousState = state;
       });
@@ -195,9 +193,7 @@ class _BlocListenerBaseState<C extends IsolateBloc<Object, S>, S>
   }
 
   void _unsubscribe() {
-    if (_subscription != null) {
-      _subscription.cancel();
-      _subscription = null;
-    }
+    _subscription?.cancel();
+    _subscription = null;
   }
 }
