@@ -31,13 +31,13 @@ mixin IsolateBlocProviderSingleChildWidget on SingleChildWidget {}
 /// );
 /// ```
 /// {@endtemplate}
-class IsolateBlocProvider<T extends IsolateBloc<Object, State>, State>
-    extends SingleChildStatelessWidget
+class IsolateBlocProvider<T extends IsolateBloc<Object, State>,
+        State extends Object> extends SingleChildStatelessWidget
     with IsolateBlocProviderSingleChildWidget {
   /// {@macro bloc_provider}
   IsolateBlocProvider({
-    Key key,
-    Widget child,
+    Key? key,
+    Widget? child,
   }) : this._(
           key: key,
           create: (_) => null,
@@ -63,9 +63,9 @@ class IsolateBlocProvider<T extends IsolateBloc<Object, State>, State>
   /// );
   /// ```
   IsolateBlocProvider.value({
-    Key key,
-    @required IsolateBlocWrapper value,
-    Widget child,
+    Key? key,
+    required IsolateBlocWrapper<Object> value,
+    Widget? child,
   }) : this._(
           key: key,
           create: (_) => value,
@@ -75,20 +75,20 @@ class IsolateBlocProvider<T extends IsolateBloc<Object, State>, State>
   /// Internal constructor responsible for creating the [IsolateBlocProvider].
   /// Used by the [IsolateBlocProvider] default and value constructors.
   IsolateBlocProvider._({
-    Key key,
-    @required Create<IsolateBlocWrapper> create,
-    Dispose<IsolateBlocWrapper<Object>> dispose,
+    Key? key,
+    required Create<IsolateBlocWrapper<Object>?> create,
+    Dispose<IsolateBlocWrapper?>? dispose,
     this.child,
   })  : _dispose = dispose,
         _create = create,
         super(key: key, child: child);
 
   /// [child] and its descendants which will have access to the `bloc`.
-  final Widget child;
+  final Widget? child;
 
-  final Create<IsolateBlocWrapper<Object>> _create;
+  final Create<IsolateBlocWrapper<Object>?> _create;
 
-  final Dispose<IsolateBlocWrapper<Object>> _dispose;
+  final Dispose<IsolateBlocWrapper?>? _dispose;
 
   /// Method that allows widgets to access a `cubit` instance as long as their
   /// `BuildContext` contains a [IsolateBlocProvider] instance.
@@ -100,10 +100,11 @@ class IsolateBlocProvider<T extends IsolateBloc<Object, State>, State>
   /// IsolateBlocProvider.of<BlocA, BlocAState>(context)
   /// ```
   static IsolateBlocWrapper<State>
-      of<T extends IsolateBloc<Object, State>, State>(BuildContext context) {
+      of<T extends IsolateBloc<Object, State>, State extends Object>(
+          BuildContext context) {
     final blocInfoHolder = _getBlocInfoHolder(context);
-    if (blocInfoHolder == null ||
-        blocInfoHolder.getWrapperByType<T, State>() == null) {
+    final blocWrapper = blocInfoHolder?.getWrapperByType<T, State>();
+    if (blocWrapper == null) {
       throw FlutterError(
         '''
         IsolateBlocProvider.of() called with a context that does not contain a IsolateBlocWrapper for $T.
@@ -115,28 +116,32 @@ class IsolateBlocProvider<T extends IsolateBloc<Object, State>, State>
         ''',
       );
     }
-    return blocInfoHolder.getWrapperByType<T, State>();
+    return blocWrapper;
   }
 
   @override
-  Widget buildWithChild(BuildContext context, Widget child) {
-    return InheritedProvider<BlocInfoHolder>(
+  Widget buildWithChild(BuildContext context, Widget? child) {
+    return InheritedProvider<BlocInfoHolder?>(
       create: (context) {
-        var blocWrapper = _create?.call(context);
+        var blocWrapper = _create.call(context);
         blocWrapper ??= createBloc<T, State>();
+        if (blocWrapper == null) {
+          print("Failed to create bloc");
+          return null;
+        }
         final blocInfoHolder = _getBlocInfoHolder(context) ?? BlocInfoHolder();
         blocInfoHolder.addBlocInfo<T>(blocWrapper);
         return blocInfoHolder;
       },
       dispose: (context, infoHolder) {
-        _dispose?.call(context, infoHolder.removeBloc<T>());
+        _dispose?.call(context, infoHolder?.removeBloc<T>());
       },
-      child: child,
       lazy: false,
+      child: child,
     );
   }
 
-  static BlocInfoHolder _getBlocInfoHolder(BuildContext context) {
+  static BlocInfoHolder? _getBlocInfoHolder(BuildContext context) {
     try {
       return Provider.of<BlocInfoHolder>(context, listen: false);
     } catch (_) {
@@ -156,7 +161,8 @@ extension IsolateBlocProviderExtension on BuildContext {
   /// ```dart
   /// BlocProvider.of<C>(context)
   /// ```
-  IsolateBlocWrapper<State>
-      isolateBloc<C extends IsolateBloc<Object, State>, State>() =>
-          IsolateBlocProvider.of<C, State>(this);
+  IsolateBlocWrapper<State> isolateBloc<C extends IsolateBloc<Object, State>,
+      State extends Object>() {
+    return IsolateBlocProvider.of<C, State>(this);
+  }
 }
