@@ -15,18 +15,50 @@ class IsolatedConnector {
 
   void _listener(ServiceEvent event) {
     if (event is IsolateBlocTransitionEvent) {
-      IsolatedBlocManager.instance
-          .blocEventReceiver(event.blocUuid, event.event);
+      final blocManager = IsolatedBlocManager.instance;
+      if (blocManager == null) {
+        print("Failed to receive event in Isolate. Bloc manager is null");
+      } else {
+        blocManager.blocEventReceiver(event.blocUuid, event.event);
+      }
     } else if (event is CreateIsolateBlocEvent) {
-      IsolatedBlocManager.instance.createBloc(event.blocType);
+      final blocManager = IsolatedBlocManager.instance;
+      if (blocManager == null) {
+        print("Failed to create IsolateBloc. Bloc manager is null");
+      } else {
+        blocManager.createBloc(event.blocType);
+      }
     } else if (event is CloseIsolateBlocEvent) {
-      IsolatedBlocManager.instance.closeBloc(event.blocUuid);
+      final blocManager = IsolatedBlocManager.instance;
+      if (blocManager == null) {
+        print("Failed to close IsolateBloc. Bloc manager is null");
+      } else {
+        blocManager.closeBloc(event.blocUuid);
+      }
     } else if (event is PlatformChannelResponseEvent) {
-      IsolatedPlatformChannelMiddleware.instance
-          .platformChannelResponse(event.id, event.data);
+      final middleware = IsolatedPlatformChannelMiddleware.instance;
+      if (middleware == null) {
+        print(
+          "Failed to receive platform channel data. "
+          "Platform channel middleware is null",
+        );
+      } else {
+        middleware.platformChannelResponse(event.id, event.data);
+      }
     } else if (event is InvokeMethodChannelEvent) {
-      IsolatedPlatformChannelMiddleware.instance
-          .handlePlatformMessage(event.channel, event.id, event.data);
+      final middleware = IsolatedPlatformChannelMiddleware.instance;
+      if (middleware == null) {
+        print(
+          "Failed to send platform channel data. "
+          "Platform channel middleware is null",
+        );
+      } else {
+        middleware.handlePlatformMessage(
+          event.channel,
+          event.id,
+          event.data,
+        );
+      }
     }
   }
 }
