@@ -23,43 +23,49 @@ void runTests() {
       "Test basic functionality such as create bloc, add event, receive state",
       () {
     test(
-        "Test that initialize function works fine and doesn't terminate process",
-        () async {
-      await testInitialize(simpleBlocInitializer);
-      expect(true, true);
-    });
+      "Test that initialize function works fine and doesn't terminate process",
+      () async {
+        await testInitialize(simpleBlocInitializer);
+        expect(true, true);
+      },
+    );
+
     test(
-        "Test that bloc successfully created, "
-        "emit it's initial state and this state received.", () async {
-      await testInitialize(simpleBlocInitializer);
-      final blocWrapper = createBloc<SimpleBloc, String>();
-      var valueReceived = false;
-      blocWrapper.listen((_) {
-        valueReceived = true;
-      });
-      // expect that next task in the event loop will be my listen function so i wait for this task.
-      // and it is really works! just try to comment this line, run test and it will fail.
-      await Future.delayed(Duration.zero);
-      expect(
-        valueReceived,
-        true,
-        reason: "Didn't receive initial state immediately",
-      );
-    });
+      "Test that bloc successfully created, "
+      "emit it's initial state and this state received.",
+      () async {
+        await testInitialize(simpleBlocInitializer);
+        final blocWrapper = createBloc<SimpleBloc, String>();
+        var valueReceived = false;
+        blocWrapper!.listen((_) {
+          valueReceived = true;
+        });
+        // expect that next task in the event loop will be my listen function so i wait for this task.
+        // and it is really works! just try to comment this line, run test and it will fail.
+        await Future.delayed(Duration.zero);
+        expect(
+          valueReceived,
+          true,
+          reason: "Didn't receive initial state immediately",
+        );
+      },
+    );
+
     test("Test that bloc emits its initial state", () async {
       await testInitialize(simpleBlocInitializer);
       final blocWrapper = createBloc<SimpleBloc, String>();
-      final initialState = await blocWrapper.first;
+      final initialState = await blocWrapper!.first;
       expect(
         initialState,
         'empty',
         reason: "First simple bloc's state must be an empty word",
       );
     });
+
     test("Test that bloc emits it's not initial state", () async {
       await testInitialize(simpleBlocInitializer);
       final blocWrapper = createBloc<SimpleBloc, String>();
-      blocWrapper.add(Object());
+      blocWrapper!.add(Object());
       final stateFuture = blocWrapper.skip(1).first;
       var stateReceived = false;
       // wait for bloc's response
@@ -74,19 +80,21 @@ void runTests() {
       await stateFuture;
       stateReceived = true;
     });
+
     test("Test that bloc emits state with data on some event", () async {
       await testInitialize(simpleBlocInitializer);
       final blocWrapper = createBloc<SimpleBloc, String>();
-      blocWrapper.add(Object());
+      blocWrapper!.add(Object());
       final state = await blocWrapper.skip(1).first;
       expect(state, 'data', reason: "Bloc's not initial state is not right");
     });
+
     test("Test that state received on every event", () async {
       await testInitialize(counterBlocInitializer);
       final blocWrapper = createBloc<CounterBloc, int>();
       const eventsCount = 4;
       var statesReceivedCount = 0;
-      blocWrapper.listen((state) {
+      blocWrapper!.listen((state) {
         statesReceivedCount++;
       });
       for (int i = 0; i < eventsCount; i++) {
@@ -101,13 +109,14 @@ void runTests() {
             "number of events",
       );
     });
+
     test("Test that bloc's state received in right order", () async {
       await testInitialize(counterBlocInitializer);
       final blocWrapper = createBloc<CounterBloc, int>();
       final expectOrder = [0, 1, 2, 1];
       // Point to the index of the next state to check
       var receivedStatePointer = 0;
-      blocWrapper
+      blocWrapper!
         ..add(true)
         ..add(true)
         ..add(false)
@@ -120,9 +129,10 @@ void runTests() {
           );
         });
     });
+
     test("Test that IsolateBlocWrapper closed when call close()", () async {
       await testInitialize(simpleBlocInitializer);
-      final blocWrapper = createBloc<SimpleBloc, String>();
+      final blocWrapper = createBloc<SimpleBloc, String>()!;
       var streamClosed = false;
       blocWrapper.listen((_) {}, onDone: () {
         streamClosed = true;
@@ -135,31 +145,33 @@ void runTests() {
 
   group("Some more complex tests", () {
     test(
-        "Test that two bloc of the same type actually are different "
-        "and emits different states", () async {
-      await testInitialize(counterBlocInitializer);
-      final bloc1 = createBloc<CounterBloc, int>();
-      final bloc2 = createBloc<CounterBloc, int>();
-      bloc1.add(true);
-      bloc1.skip(1).listen((state) {
-        expect(
-          state,
-          1,
-          reason: "Received wrong state from another bloc "
-              "or this bloc receive event for another bloc",
-        );
-      });
-      bloc2.add(false);
-      bloc2.skip(1).listen((state) {
-        expect(
-          state,
-          -1,
-          reason: "Received wrong state from another bloc "
-              "or this bloc receive event for another bloc",
-        );
-      });
-      // wait for responses from blocs
-      await Future.delayed(const Duration(milliseconds: 10));
-    });
+      "Test that two bloc of the same type actually are different "
+      "and emits different states",
+      () async {
+        await testInitialize(counterBlocInitializer);
+        final bloc1 = createBloc<CounterBloc, int>()!;
+        final bloc2 = createBloc<CounterBloc, int>()!;
+        bloc1.add(true);
+        bloc1.skip(1).listen((state) {
+          expect(
+            state,
+            1,
+            reason: "Received wrong state from another bloc "
+                "or this bloc receive event for another bloc",
+          );
+        });
+        bloc2.add(false);
+        bloc2.skip(1).listen((state) {
+          expect(
+            state,
+            -1,
+            reason: "Received wrong state from another bloc "
+                "or this bloc receive event for another bloc",
+          );
+        });
+        // wait for responses from blocs
+        await Future.delayed(const Duration(milliseconds: 10));
+      },
+    );
   });
 }
